Test: cts-cli: Use only the first line of cmd in output
Multi-line commands -- for example, those containing a multi-line string
argument -- aren't parsed correctly. All lines after the first appear in
the output (as $app), with everything after the first space (if any)
removed by sed. This caused a ton of whitespace in the output of a
couple of recently added crm_verify commands and caused the summary
"Passed" lines to print incorrectly.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>