Low: pacemaker-attrd: Drop "requesting shutdown" code
The requesting_shutdown variable was checked only by
attrd_shutting_down(), when the if_requested argument was set to true.
In that case, it returned true if either the shutting_down variable was
true or both the if_requested argument and the requesting_shutdown
variable were true.

The only caller that passed if_requested=true was
attrd_cib_updated_cb(). It did this if:
a. the alerts section was changed, or
b. the status section or nodes section was changed by an untrusted
  client.

Details:
a. Prior to f42e170, we didn't pass if_requested=true for an alerts
   section change. We started doing so as of that commit mostly for
   convenience. We decided that it seemed reasonable to ignore alert
   changes when there was a shutdown pending.

   This commit reverts to NOT ignoring alert changes due to pending
   shutdown. That seems like it might be better. I'm not sure if it's
   possible for us to land in attrd_send_attribute_alert() while a
   shutdown is requested but has not begun. If so, it would be good to
   send the correct alerts.

b. The other call with true is to avoid writing out all attributes when
   the status or nodes section changes. It's probably okay to drop the
   true there too. It was added by a1a9c54, to resolve a race condition
   where:
   * node2 left.
   * node1's controller deleted node2's transient attributes from the
     CIB.
   * node1 took over as DC and replaced the CIB.
   * node2's attribute manager was not yet actually shutting down, and
     it responded to the CIB replacement by writing out all of the
     attributes that were in its memory, including its own "shutdown"
     attribute.

   Now (as of the previous commit), node1's attribute manager would
   delete this "shutdown" attribute as part of its shutdown process. (Or
   more accurately, I think the attribute writer node will do that.)

   So if we understand correctly, the attrd_shutting_down(true)
   workaround is no longer needed.

With no more callers needing to pass true, the supporting code can go
away.

Co-Authored-By: Reid Wahl <nrwahl@protonmail.com>