Fix: controller: don't purge transient attributes on node loss
With recent changes, the attribute manager now handles it when the node
leaves the cluster, so the controller purge is redundant.

This does alter the timing somewhat, since the controller's purge
occurred when the node left the controller process group, while the
attribute manager's purge occurs when it leaves the cluster, but that
shouldn't make a significant difference.

This fixes a problem when a node's controller crashes and is respawned
while fencing is disabled. Previously, another node's controller would
remove that node's transient attributes from the CIB, but they would
remain in the attribute managers' memory. Now, the attributes are
correctly retained in the CIB in this situation.

Fixes T137
Fixes T139

Co-Authored-By: Chris Lumens <clumens@redhat.com>