Test: libcrmcommon: Add wrapper around abort() for unit tests.
This is always enabled for unit tests and does not affect when we are
building for real.  This is needed to ensure that code that aborts still
writes out code coverage numbers.  See comments for details.

Additionally, call __gcov_dump where we would call _exit() in our
pcmk__assert_* macros for the same reason.