Refactor: libcrmcommon: Rename PCMK__ENV_PHYSICAL_HOST to PCMK__META
It's not an environment variable, except in the way that all
meta-attributes become environment variables for the resource agent.

We also change from "physical_host" to "physical-host". This won't cause
any resource definition changes, because crm_meta_name() has always
converted the hyphen to an underscore.

This is a special meta-attribute that only Pacemaker should set. Users
should not be setting it; behavior would be undefined, presumably with
Pacemaker overwriting the user-configured value.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>