Refactor: libcrmcommon: New PCMK__XA_SUBT string constant
"SUBT" is so obfuscated that I considered naming it PCMK__XA_SUBTYPE
instead of aligning with the literal "subt". Unfortunately, we'll be
converting F_TYPE next, and its literal is "t". We can't use
PCMK__XA_TYPE for that because it's already defined as "type".

It would be nice to replace "t" with "type" and "subt" with "subtype"
but that means a long period of supporting both during rolling upgrades.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>