Refactor: various: Drop F_XML_TAGNAME internally
It adds a field to a few internal XML requests, but nothing checks it.
The information is redundant; it matches the element name (who'd have
thought?).

It was checked when using the heartbeat cluster layer. However, support
for heartbeat was removed in commit 55ab749 in 2.0.0, so we don't have
to worry about an older node running heartbeat during rolling upgrades.
(It would be incompatible with the newer nodes running corosync.)

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>