Refactor: libcrmcommon: New PCMK__XA_OBJECT_TYPE string constant
This is used only in the reply/output/whatever of a failed CIB create
operation. It will be available to the registered callback function, so
users do have access to it. However, the structure of the response is
not documented or guaranteed, so users shouldn't rely on a particular
format.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>