Refactor: libcib: Use object-type in CIB create failure reply
...instead of object_type.

This affects the data received by a client's callback function. We don't
document the reply format or make any guarantees about it, so we don't
need to maintain compatibility by using the same attribute names
forever.

This allows us to have only PCMK_XA_OBJECT_TYPE, instead of both
PCMK_XA_OBJECT_TYPE (public, with hyphen in the literal) and
PCMK__XA_OBJECT_TYPE (internal, with underscore in the literal).

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>