Refactor: libcrmcommon: New enum pcmk__opt_context
At first this will be used only for backward compatibility with metadata
commands for pacemaker-based, pacemaker-controld, and
pacemaker-schedulerd. It may be extended later for meta-attribute
contexts or similar.

The idea is that we'll consolidate all cluster options into a single
table and use these enum values as filters for getting daemon metadata.

We won't need a pcmk__opt_context_fenced, because its metadata consists
of stonith instance attributes, not cluster options. Those instance
attributes will be stored in a separate array from the cluster options.

Ref T746

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>