Low: libcib: Do not check CIB feature set for files in cib_perform_op.
This is related to the previous feature for transferring schema files to
older remote nodes.  In that case, the newer schema files may also have
a newer feature set than the node supports, so the transferred files are
still not usable.

However, the feature set only matters for the scheduler, not for most
command line tools (obviously, crm_simulate would still care).  So in
those cases, we can just disable the feature set check if the CIB was
read in from a file.  For the scheduler, the check is still performed as
part of cluster_status.