Refactor: libcrmcommon: New filter arg to pcmk__format_option_metadata()
Now each cluster option is in exactly one daemon's metadata. The four
options that were previously in the metadata of both the controller and
the scheduler are now only in the scheduler's metadata.

All daemons still have access to all the options they use.

All function calls in daemons still use the local options arrays rather
than the one in libcrmcommon. That will change in upcoming commits.

Closes T746

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>