Fix: pacemaker-attrd: wipe CIB along with memory
Previously, when the attribute manager purged a node, it would purge the node's
transient attributes only from memory, and assumed the controller would purge
them from the CIB.

Now, the writer will purge them from the CIB as well. The attributes will be
initially kept in memory and marked as removed, then when the CIB confirmation
is received, they will actually be removed. (The attrd protocol version will be
bumped in a separate commit.)

This fixes a variety of timing issues when multiple nodes are shutting down,
including the attribute writer, which leaves before some other node. When that
other node leaves the controller process group, the DC wipes that node's
attributes from the CIB (or all other nodes do if the DC is the leaving node).
If a new writer (possibly even the node itself) is elected before the node's
attribute manager leaves the cluster layer, it will write the attributes back
to the CIB. Once the other node leaves the cluster layer, all attribute
managers remove its attributes from memory, but they are now "stuck" in the
CIB.

With the new approach, once the other node leaves the cluster, the next elected
writer will wipe its attributes from the CIB (whether the writer is elected
before or after that the node leaves).

This also fixes a problem when a node's controller crashes and is respawned
(with fencing disabled). Previously, another node's controller would remove
that node's transient attributes from the CIB, but they would remain in the
attribute managers' memory. Now, the attributes are correctly retained in the
CIB in this situation.

This causes a known regression. Currently, the controller purges attributes
from the CIB for leaving DCs only if they are at version 1.1.13 or later,
because earlier DCs will otherwise get fenced after a clean shutdown. Since the
attribute manager doesn't know the DC or its version, the attributes will now
always be wiped. However, this seems acceptable since the problem can occur
only in the highly unlikely situation of a rolling upgrade from supported
versions 1.1.11 or 1.1.12, and the rolling upgrade will still succeed without
any negative impact on resources.

As of this commit, the controller still erases the attributes from the CIB when
the node leaves the controller process group, which is redundant but doesn't
hurt anything and will be corrected in an upcoming commit.

Fixes T137
Fixes T138
Fixes T139