Refactor: tools: Rename node-list to simple-node-list in crm_node.
Formatted output messages are supposed to be unique throughout
pacemaker.  At some point, we gained a node-list message in both
crm_node and in libpe.  It doesn't matter which came first, but the
libpe one is now also in use by libpacemaker, while the crm_node one
only exists in that tool.  So rename it to be unique.