Fix: tools: crm_attribute emits garbage for --node localhost or auto
This happens because pcmk__node_attr_target() returns its argument if
its argument is NULL, "auto", or "localhost" and no relevant environment
variables are found. Then crm_attribute frees the return value, makes a
copy of it, and assigns it back to options.dest_uname.

The fix is to check whether the return value is equal to the argument.

Fixes RHEL-23065

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>