Refactor: libcrmcommon: New pcmk__xml_needs_escape()
To avoid unnecessary allocations.

We skip this in add_desc() simply because that logic is going away in an
upcoming commit. We'll be allocating an XML text node instead, which we
escape in the constructor.

We skip this in services_lsb.c because of the number of variables
required and the infrequency of the calls.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>