Test: cts-cli: Add ticket querying output to cts-cli.
This is in preparation for moving these commands to formatted output so
I can verify that I haven't changed the text output.  Note that this
includes XML output, which is not yet supported in crm_ticket aside from
the command line option that enables it, so the XML output in this patch
will be wrong.

Further note that this includes failure output from xmllint, which is to
be expected given that the XML output is not hooked up yet.