Feature: tools: Use the ticket output message in crm_ticket.
There was already a ticket message in pe_output.c, but its output format
is not quite the same as what crm_ticket was already doing.  The text
format is especially tricky because it was in use in both places.  I've
tried to preserve backwards compatability while condensing them into a
single function.

XML was easier because crm_ticket didn't already provide that output.
HTML can be removed as a separate function because it's now the same as
text.