Refactor: scheduler: The ticket-list message should take a hash table...
instead of the scheduler.  We don't need the scheduler for any reason
aside from getting at the tickets hash table.

The ticket message always calls out->list_item, which means it requires
that out->begin_list is called before and out->end_list is called after.
Not doing so will result in a segfault.  We can enforce this by always
going through the ticket-list message.

However, when we want to output a single ticket via ticket-list, we need
to construct a temporary object.  The options are a scheduler (no
thanks), a list (which would mean converting the scheduler ticket hash
table to a list first), or a singleton hash table.  The latter seems
like the easiest choice.