Refactor: libcrmcommon,libpe_rules: move operation evaluation to libcrmcommon
Move the functionality of pe__eval_op_expr() to a new function
pcmk__evaluate_op_expression() in libcrmcommon.

The new function uses pcmk_rule_input_t instead of pe_rule_eval_data_t, and
returns a standard code instead of gboolean. The old function is kept (for now)
as a wrapper that maps the old struct to the new.

No code changes other than the minimum necessary to use the new struct and
return value.