Refactor: libpacemaker: Clean up pcmk_verify.c somewhat
Account for leading whitespace in pcmk__parse_cib(). The extra "first"
variable exists instead of incrementing the cib_source pointer, just for
the edge case where a file name begins with whitespace.

It would be nice to take an enum for the type of CIB source once we can
change the public API arguments for pcmk_verify(). The current way seems
a bit error-prone; also, another edge case is that a file name can begin
with '<'.

Finally, remove trailing whitespace in pcmk_verify().

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>