API: schemas: Add a schema for crm_ticket.
This first breakes the <ticket> schema out of crm_mon into its own file,
and then adds a reference to that to crm_mon as well as creates a schema
for the crm_ticket command.

This is in preparation for adding XML formatted output to crm_ticket.
At the moment, the schema is only present to make sure that as things
get converted, the output is correct.