Fix: libcrmcommon: pcmk__xml_parse_file() recovery works for stdin
Previously, if pcmk__xml_parse_file() was called with an argument of
NULL or "-", we told libxml2 to read from stdin's file descriptor and
parse XML from it. If this failed, we tried the same thing again.

The problem is that the data from stdin has already been read and
discarded on the first attempt. So the second attempt with
XML_PARSE_RECOVER finds an empty document.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>