Refactor: various: Replace find_xml_node() with pcmk__xe_match_name()
The only significant difference is that find_xml_node() can log a
warning if must_find is TRUE.

This isn't particularly useful. In the handful of places where must_find
was set to TRUE internally, we now ensure the caller logs an error or
CRM_CHECKs the result. (Exception: controld_record_action_event(). It's
not clear to me whether we should CRM_CHECK(action_rsc != NULL), so I
left it as a simple return.)

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>