Refactor: libcrmcommon: Accept NULL parent in pcmk__xe_match()
Here we give up some mild sanity checking of existing use cases so that
we can replace find_xml_node() in an upcoming commit. find_xml_node() is
used in some cases where parent may be NULL, and it's not worth tracking
down every case and avoiding it at the caller side.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>