Refactor: libcrmcommon,libpe_rules: move ruleset evaluation to libcrmcommon
Move the functionality of pe_eval_rules() to a new function
pcmk__evaluate_rules() in libcrmcommon. It is internal because it is
needed only for deprecated functionality that will be dropped soon.

The new function uses pcmk_rule_input_t instead of pe_rule_eval_data_t, and
returns a standard Pacemaker return code. The old function is kept (for now)
as a wrapper that maps the old struct to the new.