Refactor: libcrmcommon: New pcmk__output_text_{get,set}_fancy()
Use set instead of forcing --text-fancy in CLI tools. Create a new
priv->fancy member that we'll now check in addition to the fancy static
variable, and that the new function will set to true or false.

The fancy static variable has another shortcoming besides being tied to
a CLI option: it applies to **all** text output objects in a given
program. An upcoming commit will hide and deprecate --text-fancy, and
we'll drop it and the static variable in a future release.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>