Refactor: libcrmcommon: Get rid of --xml-legacy.
Even though this was hidden, it could still be used if you knew it was
there.  This option shouldn't be exposed to users at all since it can
change the XML format in ways that wouldn't pass validation for almost
all tools.

Instead, the only user (crm_mon) can just enable legacy XML output mode
with a private library function.