Refactor: libcrmcommon: Check simple_list in both locations.
This is redundant for the moment, but it's setting up for when
simple_list is no longer a global variable.  Also, rename the new
location to something more accurate.  No one can remember what a "simple
list" is.

This also creates some goofy conditionals, but they'll be cleaned up
when I get rid of the simple_list global.