Refactor: libcrmcommon: Get rid of unnecessary output args structs.
We've had the ability to pass NULL for the option structs for a long
time, but we just haven't been using it.  Now that all the XML options
are gone, we might as well remove that struct and then all the others
that are empty.

Eventually, all of this will go.