Log: libcrmcommon: Drop "Couldn't parse XML" error messages
Because it's at least partially redundant.

pcmk__log_xmllib_err() gives us some info about the parse failure. We're
probably losing the potential to show information about exactly **what**
we failed to parse... I'm not sure that pcmk__log_xmllib_err() gives us
the file name if we fail to parse XML from a file, and I don't believe
it shows the **full** input if we fail to parse XML from a stream or a
string.

Signed-off-by: Reid Wahl <nrwahl@protonmail.com>