Refactor: test: Add CIB-related unit test group setup code.
Any group of unit tests that deals with the CIB needs to do a little
setup before running the tests.  For the moment, this is not very much
code but it could grow over time and as we add more tests, we don't want
to keep copying this stuff around everywhere.