Low: libcrmcommon: Free text/curses private list data.
Whenever we're done with the private list data in the text and curses
formatters, make sure to clean it up.  Typically, this should just be
when end_list is called.  However, if for some reason the program ends
before everything that was pushed into the queue is popped with
end_list, also clean up there too.

Ref T704