Fix: fencer: ask an online target of watchdog fencing to acknowledge an `on` action
... rather than waiting on a watchdog fencing timer and assuming a
success.

AFAICT an `on` action shouldn't go into
`check_watchdog_fencing_and_wait()` at all, since it could not be
technically achieved via a watchdog anyway, so there's no need to wait
on that timer. For watchdog fencing, an `on` action should basically
succeed as long as the fencing target is online, but probably should
just fail if the target is offline.

In here, since the peer e.g. the target node is online and claiming it
can perform an `on` action, with the changed logic, we basically allow
it to do so the usual way with the usual timer. Then the target will
perform the "unfencing", and it'll be expected to remain online and
reply to us. The target will actually fake/acknowledge the success
for us in stonith_device_execute().