Fix: fencer: make an `on` action for an offline target of watchdog fencing fail
... rather than waiting on a watchdog fencing timer and assuming a
success.

AFAICT an `on` action shouldn't go into
`check_watchdog_fencing_and_wait()` at all, since it could not be
technically achieved via a watchdog anyway, so there's no need to wait
on that timer. For watchdog fencing, an `on` action should basically
succeed as long as the fencing target is online, but probably should
just fail if the target is offline.

In here, basically the fencing target is offline and no one claims to be
able to unfence it. With the changed logic, we make it fail.